backend/ai_service/requirements.txt
# India Biodiversity AI Service - Dependencies
# Install with: pip install -r requirements.txt

# Core FastAPI web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data handling and validation
Pillow==10.2.0
numpy==1.26.4
pydantic==2.6.1
pydantic-settings==2.2.1

# HTTP client (for health checks and external APIs)
httpx==0.27.0

# ─── Optional: ML Inference ────────────────────────────────────────────────
# Install tensorflow for actual ML inference with MobileNetV2.
# If not installed, the service falls back to a rule-based classifier.
# For CPU-only systems, use: pip install tensorflow-cpu
# For lighter weight, consider: pip install tf-keras
#
tensorflow==2.16.1          # Full TensorFlow with GPU support
# tensorflow-cpu==2.16.1     # TensorFlow CPU-only (smaller download)
# tf-keras==2.16.1           # Alternative lighter Keras wrapper
#
# To install CPU-only version:
#   pip install tensorflow-cpu==2.16.1
#
# Comment out the line below if installing tensorflow-cpu instead
tensorflow==2.16.1